[MIG] Split module account_financial_discount to account_financial_discount_reconciliation_widget_oca
