--- account_financial_discount/views/payment_receipt.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template
-        id="report_payment_receipt_document"
-        inherit_id="account.report_payment_receipt_document"
-    >
-        <xpath expr="//table/thead/tr/th[3]" position="after">
-            <th class="text-right">
-                <span>Discount</span>
-            </th>
-        </xpath>
-        <xpath expr="//table/tbody/tr/td[3]" position="after">
-            <td class="text-right">
-                <span
-                    t-esc="o.get_discount_for_invoices(inv)"
-                    t-options="{'widget': 'monetary', 'display_currency': inv.currency_id}"
-                />
-            </td>
-        </xpath>
-        <xpath
-            expr="//span[@t-esc='o._get_invoice_payment_amount(inv)']"
-            position="attributes"
-        >
-            <attribute
-                name="t-esc"
-            >o._get_invoice_payment_amount(inv) + o.get_discount_for_invoices(inv)</attribute>
-        </xpath>
-        <xpath expr="//tbody" position="after">
-            <t t-set="invoices" t-value="o.reconciled_invoice_ids" />
-            <tfooter t-if="invoices">
-                <tr style="font-weight: bold">
-                    <t t-set="curr" t-value="invoices[0].currency_id" />
-                    <t
-                        t-set="total_paid"
-                        t-value="sum(o._get_invoice_payment_amount(inv) for inv in invoices)"
-                    />
-                    <td />
-                    <td />
-                    <td />
-                    <td />
-                    <td class="text-right">
-                        <span
-                            t-esc="o.get_discount_for_invoices(invoices)"
-                            t-options="{'widget': 'monetary', 'display_currency': curr}"
-                        />
-                    </td>
-                    <td class="text-right">
-                        <span
-                            t-esc="sum(invoices.mapped('amount_total'))"
-                            t-options="{'widget': 'monetary', 'display_currency': curr}"
-                        />
-                    </td>
-                    <td class="text-right">
-                        <span
-                            t-esc="total_paid + o.get_discount_for_invoices(invoices)"
-                            t-options="{'widget': 'monetary', 'display_currency': curr}"
-                        />
-                    </td>
-                    <td class="text-right">
-                        <span
-                            t-esc="sum(invoices.mapped('amount_residual'))"
-                            t-options="{'widget': 'monetary', 'display_currency': curr}"
-                        />
-                    </td>
-                </tr>
-            </tfooter>
-        </xpath>
-    </template>
-</odoo>
