--- account_financial_discount/views/payment_term_form.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="view_payment_term_form" model="ir.ui.view">
-        <field name="name">account.payment.term.form</field>
-        <field name="model">account.payment.term</field>
-        <field name="inherit_id" ref="account.view_payment_term_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//separator[1]" position="after">
-                <group>
-                    <field name="days_discount" />
-                    <field name="percent_discount" />
-                </group>
-            </xpath>
-        </field>
-    </record>
-</odoo>
