+++ account_financial_discount_reconciliation_widget_oca/models/__init__.py
@@ -1,0 +1,2 @@
+from . import account_reconcile_model
+from . import reconciliation_widget
