+++ account_financial_discount_reconciliation_widget_oca/readme/CONFIGURE.rst
@@ -1,0 +1,6 @@
+Go to Accounting > Configuration > Reconciliation models, to define a
+reconciliation model applying financial discount.
+
+* Set write-off accounts for expenses and revenues
+* Set a label for write-off line
+* Set a tolerance to apply discount (e.g. 0.05 for 5 cts)
