+++ account_financial_discount_reconciliation_widget_oca/static/src/xml/reconciliation_templates.xml
