+++ account_financial_discount_reconciliation_widget_oca/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_financial_discount_reconciliation
